ReadMe
My old man's Servicing department is collapsing from underneath itself,
if you know what i mean.. Not because of no business, but because  we
have too much work! Our contracts are not probably recorded, so many
people come to us to service/repair UPS it remarkable.. But our techs
and admin staff cant keep up with the demand of getting every machine
fixed, within the five days we say it takes to analyse, quote, get the
"go-ahead", then carry out the repair, should i say, that really means
to finish the repair. Get it marked as "Tech completed repair". The
other problem is the field and what happens in the reality of UPS field
work.. I have been there, fixed lots of machines and although i enjoyed
the work, it can be very stressful when working with power supply, the
bloodline of any enterprise... big or small. It's stressful, because,
not only is it extremely dangerous, if a tech drops that load on the
UPS, there is everything down china! I'm talking servers, lights, CCTV,
elevators, well anything that is connected to the UPS . So naturally our
customers get very sensitive when our boys pop around to come play games
with their back-up power systems. perfectly understandable.. and it is
the real reason why the market is very lucrative and busy.

Now I have one peeve about the nature and politics of our company,
unfortunately such thing are a vital requirement to keep the money
coming in and getting profit off our expenses. Salesman in my company
don't make life very easy for our already crippling Service department..
Fucking hate it the bastards, but I love them with all my heart, because
the keep bringing in the dolla! I come from working in the Matla
workshop & factory, where we built our fucking awesome designs of our
machines! But time changes, and Matla TPM would be the last South
African UPS manufacturer to stop operations in production, due to the
expense of building our machines and the recession. It is cheaper to buy
machines abroad than to manufacture your designs.. Everything comes
abroad anyway!  Anyway this program "LogTrack" is my new job at work and
the project will never end, it will consume my entire working career. I
have no desire to have a boss, other than my dad ... :-P . I now sit
with the challenge to get my service department jacked up like the guys
@ APC or MGE or Emerson, etc. So it all starts here.

Note* MatlaTech-DigiWorx is not a company, it is a development
department for software and web applications solutions for UPS Direct
(Pty) Ltd

Sean Brunton
Servicing Department